Tackat says we can't release a word process without a "word count" feature (especially for reviewers ;-). Also fixed some gcc3 warnings, simplified KWFormulaFrameSet constructors.
svn path=/trunk/koffice/; revision=92856

+++ kword/kwtextparag.cc
@@ -865,7 +865,7 @@ KWParagLayout::KWParagLayout( QDomElement & parentElem, KWDocument *doc )
             alignment = flow2align[flow.toInt()];
         } else {
             flow = element.attribute( "align" ); // KWord-1.0 DTD
-            alignment = flow=="right" ? Qt::AlignRight : flow=="center" ? Qt::AlignCenter : flow=="justify" ? Qt3::AlignJustify : Qt::AlignLeft;
+            alignment = flow=="right" ? (int)Qt::AlignRight : flow=="center" ? (int)Qt::AlignCenter : flow=="justify" ? (int)Qt3::AlignJustify : (int)Qt::AlignLeft;
         }
     }
